refactor: header
-Refactor gameplay settings modal logic into a UI component.

-Refactor all header menu buttons into a header menu UI component.
